+++ HotaMechanics/combat_hero.h
@@ -23,8 +23,8 @@ namespace HotaMechanics {
 		// complex getters ---------------
 		const std::vector<const CombatUnit*> getUnits() const;
 		const std::vector<const CombatUnit*> getUnitsPtrs();
-		const int getUnitId(const CombatUnit& _unit) const;
-		const int getGlobalUnitId(const CombatUnit& _unit) const;
+		const int16_t getUnitId(const CombatUnit& _unit) const;
+		const int16_t getGlobalUnitId(const CombatUnit& _unit) const;
 		// -------------------------------
 
 		// check hero state --------------
